Fixed a casting problem with the recent change to use qRound.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@7856 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/configureSO.cpp
@@ -511,5 +511,5 @@ void configureSO::sEditCreditMemoWatermark()
 
 void configureSO::sEditCreditLimit()
 {
-  _creditLimit->setDouble(qRound(_creditLimit->toDouble()));
+  _creditLimit->setDouble((double)qRound(_creditLimit->toDouble()));
 }
